Made `Models` folder and extract there all models from the `MovieQuizViewController` - made QuizQuestion.swift file in Models for the `QuizQuestion` struct - made QuizResultsViewModel.swift file in Models for the `QuizResultsViewModel` struct - made QuizStepViewModel.swift file in Models for the `QuizStepViewModel` struct
